move harmonize_usage
